Update UIControls with new features and fixes

+++ UIControls/src/ConstDef.cpp
@@ -6,6 +6,6 @@ const std::filesystem::path ConstDef::workforldPath = std::filesystem::path(".")
 const std::filesystem::path ConstDef::pathPrefix = std::filesystem::path(".");     // 资源路径前缀
 
 // 按钮相关常量
-const int ConstDef::BUTTON_CAPTION_SIZE = 16;
-const int ConstDef::WINDOW_TITLE_HEIGHT = 30;
-const int ConstDef::FONT_MARGIN = 10;
+const float ConstDef::BUTTON_CAPTION_SIZE = 16;
+const float ConstDef::WINDOW_TITLE_HEIGHT = 30;
+const float ConstDef::FONT_MARGIN = 10;
